add flows to base Worker trait and implement a basic function style version of it
